requirements_cpu.txt
torch==2.1.2+cpu
torchvision==0.16.2+cpu
torchaudio==2.1.2+cpu
tqdm>=4.66.1
matplotlib>=3.1.0
Pillow>=7.0.0
